requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI models
anthropic==0.5.0
openai==1.3.7
tavily-python>=0.3.0

# LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
chromadb==0.4.22
tiktoken==0.5.2

# Database
motor==3.3.2
sqlalchemy==2.0.23
alembic==1.12.1

# Web and async
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
websockets==12.0

# Security
bcrypt==4.0.1
cryptography==41.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pytest-cov==4.0.0

# Web scraping
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests-html==0.10.0
fake-useragent==1.4.0

# Optional: Uncomment for production
# gunicorn==21.2.0
# psycopg2-binary==2.9.7  # For PostgreSQL
# pymongo==4.6.0  # For MongoDB if needed
